\name{italymapping-package}
\alias{italymapping}
\alias{italymapping-package}
\docType{package}
\title{Italy Mapping}
\description{Offline relief maps of Italy using bundled national GIS data, English labels, regional and municipal boundaries, and an OpenStreetMap river snapshot.}
\author{Marco Aurelio Ramirez Mauricio (2026)}
\seealso{\code{\link{italy_example}}, \code{\link{plot_italy_map}}, \code{\link{italy_regions}}, \code{\link{save_italy_map}}}
